Added a preliminary implementation to merge clusters.  The cluster merging feature has been implmented as a new type of cluster-maker. The merging cluster maker extends the features of MST-based clsuter maker and then merges the generated clusters.  Similar clusters are detected by comparing sample subset of reads from the two clusters. The number of reads to compare and size of clusters to be merged are specified as command-line arguments.
Currently the merging seems to be working correctly for 1-process
runs.  The parallel run seems to have a bug that needs to be fixed.

Here is an example of how the merging cluster-maker can be used:

$ cd ~/research/peace/tools/ha_1500
$ ../../src/peace --heuristics primes tv --fastaFiles HA_500.fa --clusterMaker merging

+++ src/clustering/PrimesHeuristic.cpp
@@ -64,7 +64,7 @@ PrimesHeuristic::addCommandLineArguments(ArgParser& argParser) {
         {"--pri-heur-cg", "Prime value for C/G in primes heuristic",
          &cgPrime, ArgParser::INTEGER},
         {"--pri-heur-thresh", "Distance threshold override for similarity",
-         &distThresh, ArgParser::INTEGER},
+         &distThresh, ArgParser::LONG},
         {"--pri-heur-topN", "Restrict to top-n reads",
          &topN, ArgParser::INTEGER},        
         {"", "", NULL, ArgParser::INVALID}
